GEODE-5803 geode-core:distributedTest does not need test or integrationTest
Move refactors only
Gradle changes to trigger
