+++ geode-core/src/integrationTest/java/org/apache/geode/pdx/JSONFormatterJUnitTest.java
@@ -25,7 +25,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.json.JSONException;
-import org.json.JSONObject;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -75,7 +74,7 @@ public class JSONFormatterJUnitTest {
     PdxInstance pi = (PdxInstance) receivedObject;
     String json = JSONFormatter.toJSON(pi);
 
-    verifyJsonWithJavaObject(json, actualTestObject);
+    JSONFormatVerifyUtility.verifyJsonWithJavaObject(json, actualTestObject);
   }
 
   @Test
@@ -145,56 +144,6 @@ public class JSONFormatterJUnitTest {
 
     assertEquals("actualTestObject and expectedTestObject should be equal", expectedTestObject,
         actualTestObject);
-  }
-
-  static void verifyJsonWithJavaObject(String json, TestObjectForJSONFormatter testObject) {
-    JSONObject jsonObject = new JSONObject(json);
-
-    // Testcase-1: Validate json string against the pdxInstance.
-    // validation for primitive types
-    assertEquals("VerifyPdxInstanceToJson: Int type values are not matched", testObject.getP_int(),
-        jsonObject.getInt(testObject.getP_intFN()));
-    assertEquals("VerifyPdxInstanceToJson: long type values are not matched",
-        testObject.getP_long(), jsonObject.getLong(testObject.getP_longFN()));
-
-    // validation for wrapper types
-    assertEquals("VerifyPdxInstanceToJson: Boolean type values are not matched",
-        testObject.getW_bool().booleanValue(), jsonObject.getBoolean(testObject.getW_boolFN()));
-    assertEquals("VerifyPdxInstanceToJson: Float type values are not matched",
-        testObject.getW_double().doubleValue(), jsonObject.getDouble(testObject.getW_doubleFN()),
-        0);
-    assertEquals("VerifyPdxInstanceToJson: bigDec type values are not matched",
-        testObject.getW_bigDec().longValue(), jsonObject.getLong(testObject.getW_bigDecFN()));
-
-    // vlidation for array types
-    assertEquals("VerifyPdxInstanceToJson: Byte[] type values are not matched",
-        (int) testObject.getW_byteArray()[1],
-        jsonObject.getJSONArray(testObject.getW_byteArrayFN()).getInt(1));
-    assertEquals("VerifyPdxInstanceToJson: Double[] type values are not matched",
-        testObject.getW_doubleArray()[0],
-        jsonObject.getJSONArray(testObject.getW_doubleArrayFN()).getDouble(0), 0);
-    assertEquals("VerifyPdxInstanceToJson: String[] type values are not matched",
-        testObject.getW_strArray()[2],
-        jsonObject.getJSONArray(testObject.getW_strArrayFN()).getString(2));
-
-    // validation for collection types
-    assertEquals("VerifyPdxInstanceToJson: list type values are not matched",
-        testObject.getC_list().get(0),
-        jsonObject.getJSONArray(testObject.getC_listFN()).getString(0));
-
-    assertEquals("VerifyPdxInstanceToJson: stack type values are not matched",
-        testObject.getC_stack().get(2),
-        jsonObject.getJSONArray(testObject.getC_stackFN()).getString(2));
-
-    // validation for Map
-    assertEquals("VerifyPdxInstanceToJson: Map type values are not matched",
-        testObject.getM_empByCity().get("Ahmedabad").get(0).getFname(),
-        jsonObject.getJSONObject(testObject.getM_empByCityFN()).getJSONArray("Ahmedabad")
-            .getJSONObject(0).getString("fname"));
-
-    // validation Enum
-    assertEquals("VerifyPdxInstanceToJson: Enum type values are not matched",
-        testObject.getDay().toString(), jsonObject.getString(testObject.getDayFN()));
   }
 
   /**
